refactor(*): Create SceneDiffFlameGraph
